flask_app/requirements.txt
Flask
celery[redis]
gunicorn
Pillow